refactors: move constants to a separate file
